Align RSA interrupt clear register/field name for S2/S3 with other chips
