+++ esp32s2/svd/patches/esp32s2.yaml
@@ -345,9 +345,15 @@ USB0:
 
 RSA:
   _modify:
+    CLEAR_INTERRUPT:
+      name: INT_CLR
     INTERRUPT_ENA:
       name: INT_ENA
 
+  INT_CLR:
+    _modify:
+      CLEAR_INTERRUPT:
+        name: INT_CLR
   INT_ENA:
     _modify:
       INTERRUPT_ENA:
